[fix] create_default_tableメソッドを修正
