+++ db/migrate/20220923164123_add_default_data_to_tables.rb
@@ -1,0 +1,5 @@
+class AddDefaultDataToTables < ActiveRecord::Migration[6.1]
+  def change
+    add_column :tables, :default_data, :boolean, null: false, default: false
+  end
+end
