+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2022_08_16_113848) do
+ActiveRecord::Schema.define(version: 2022_09_23_164123) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -404,6 +404,7 @@ ActiveRecord::Schema.define(version: 2022_08_16_113848) do
     t.datetime "updated_at", precision: 6, null: false
     t.string "title"
     t.integer "user_id"
+    t.boolean "default_data", default: false, null: false
   end
 
   create_table "taxons", force: :cascade do |t|
