+++ spec/requests/top_spec.rb
@@ -3,6 +3,8 @@ require 'rails_helper'
 RSpec.describe 'Top', type: :request do
   before do
     @user = create(:user)
+    janitor = create(:janitor)
+    create(:sample_table, user: janitor)
   end
 
   describe '#top' do
